refactor: Move some functions and interfaces around to avoid circular dependencies between FilesChannel, FileCalculations, and formula-lib
