+++ src/aux-common/Files/FilesChannel.ts
@@ -1,18 +1,17 @@
-import { sortBy, flatMap, mapValues } from 'lodash';
-import { File, Object, PartialFile, PrecalculatedFile } from './File';
+import { sortBy, flatMap, mapValues, sortedIndexBy, values } from 'lodash';
+import { File, Object, PartialFile, PrecalculatedFile, FileTags } from './File';
 import {
-    createCalculationContext,
-    FileCalculationContext,
     calculateFileValue,
     getActiveObjects,
     filtersMatchingArguments,
     calculateFormulaValue,
     isFile,
     isDestroyable,
-    FileSandboxContext,
+    objectsAtContextGridPosition,
 } from './FileCalculations';
+import { FileCalculationContext, FileSandboxContext } from './FileContext';
 import { merge as mergeObj } from '../utils';
-import formulaLib, {
+import {
     setActions,
     getActions,
     setFileState,
@@ -20,7 +19,11 @@ import formulaLib, {
     getCalculationContext,
     getUserId,
     getFileState,
-} from '../Formulas/formula-lib';
+} from '../Formulas/formula-lib-globals';
+import formulaLib from '../Formulas/formula-lib';
+import SandboxInterface, { FilterFunction } from '../Formulas/SandboxInterface';
+import { SandboxLibrary, Sandbox } from '../Formulas/Sandbox';
+import uuid from 'uuid/v4';
 
 /**
  * Defines an interface for the state that an AUX file can contain.
@@ -67,6 +70,63 @@ interface FileChanges {
         changedTags: string[];
         newValues: string[];
     };
+}
+
+/**
+ * Creates a new file calculation context.
+ * @param objects The objects that should be included in the context.
+ * @param lib The library JavaScript that should be used.
+ */
+export function createCalculationContext(
+    objects: Object[],
+    userId: string = null,
+    lib: SandboxLibrary = formulaLib
+): FileSandboxContext {
+    const context = {
+        sandbox: new Sandbox(lib),
+        objects: objects,
+    };
+    context.sandbox.interface = new SandboxInterfaceImpl(context, userId);
+    return context;
+}
+
+export function createPrecalculatedContext(
+    objects: PrecalculatedFile[]
+): FileCalculationContext {
+    const context = {
+        objects: objects,
+    };
+    return context;
+}
+
+/**
+ * Creates a new file calculation context from the given files state.
+ * @param state The state to use.
+ * @param includeDestroyed Whether to include destroyed files in the context.
+ */
+export function createCalculationContextFromState(
+    state: FilesState,
+    includeDestroyed: boolean = false
+) {
+    const objects = includeDestroyed ? values(state) : getActiveObjects(state);
+    return createCalculationContext(objects);
+}
+
+/**
+ * Executes the given formula on the given file state and returns the results.
+ * @param formula The formula to run.
+ * @param state The file state to use.
+ * @param options The options.
+ */
+export function searchFileState(
+    formula: string,
+    state: FilesState,
+    { includeDestroyed }: { includeDestroyed?: boolean } = {}
+) {
+    includeDestroyed = includeDestroyed || false;
+    const context = createCalculationContextFromState(state, includeDestroyed);
+    const result = calculateFormulaValue(context, formula);
+    return result;
 }
 
 /**
@@ -834,4 +894,135 @@ export function setForcedOffline(offline: boolean): SetForcedOfflineEvent {
     };
 }
 
-// export
+class SandboxInterfaceImpl implements SandboxInterface {
+    private _userId: string;
+    objects: Object[];
+    context: FileCalculationContext;
+
+    private _fileMap: Map<string, FileTags>;
+
+    constructor(context: FileCalculationContext, userId: string) {
+        this.objects = sortBy(context.objects, 'id');
+        this.context = context;
+        this._userId = userId;
+        this._fileMap = new Map();
+    }
+
+    /**
+     * Adds the given file to the calculation context and returns a proxy for it.
+     * @param file The file to add.
+     */
+    addFile(file: File): File {
+        const index = sortedIndexBy(this.objects, file, f => f.id);
+        this.objects.splice(index, 0, file);
+        return file;
+    }
+
+    listTagValues(tag: string, filter?: FilterFunction, extras?: any) {
+        const tags = this.objects
+            .map(o => this._calculateValue(o, tag))
+            .filter(t => hasValue(t));
+        const filtered = this._filterValues(tags, filter);
+        return filtered;
+    }
+
+    listObjectsWithTag(tag: string, filter?: FilterFunction, extras?: any) {
+        const objs = this.objects.filter(o =>
+            hasValue(this._calculateValue(o, tag))
+        );
+        const filtered = this._filterObjects(objs, filter, tag);
+        return filtered;
+    }
+
+    list(obj: any, context: string) {
+        if (!context) {
+            return [];
+        }
+        const x: number = obj[`${context}.x`];
+        const y: number = obj[`${context}.y`];
+
+        if (typeof x !== 'number' || typeof y !== 'number') {
+            return [];
+        }
+
+        const objs = objectsAtContextGridPosition(this.context, context, {
+            x,
+            y,
+        });
+        return objs;
+    }
+
+    uuid(): string {
+        return uuid();
+    }
+
+    userId(): string {
+        return this._userId;
+    }
+
+    getTag(file: File, tag: string): any {
+        const tags = this._getFileTags(file.id);
+        if (tags.hasOwnProperty(tag)) {
+            return tags[tag];
+        }
+        return calculateFileValue(this.context, file, tag);
+    }
+
+    setTag(file: File, tag: string, value: any): any {
+        const tags = this._getFileTags(file.id);
+        tags[tag] = value;
+        return value;
+    }
+
+    getFileUpdates(): FileUpdatedEvent[] {
+        const files = [...this._fileMap.entries()];
+        const updates = files
+            .filter(f => {
+                return Object.keys(f[1]).length > 0;
+            })
+            .map(f =>
+                fileUpdated(f[0], {
+                    tags: f[1],
+                })
+            );
+
+        return sortBy(updates, u => u.id);
+    }
+
+    private _filterValues(values: any[], filter: FilterFunction) {
+        if (filter) {
+            if (typeof filter === 'function') {
+                return values.filter(filter);
+            } else {
+                return values.filter(t => t === filter);
+            }
+        } else {
+            return values;
+        }
+    }
+
+    private _filterObjects(objs: any[], filter: FilterFunction, tag: string) {
+        if (filter) {
+            if (typeof filter === 'function') {
+                return objs.filter(o => filter(this._calculateValue(o, tag)));
+            } else {
+                return objs.filter(o => this._calculateValue(o, tag) == filter);
+            }
+        } else {
+            return objs;
+        }
+    }
+
+    private _calculateValue(object: any, tag: string) {
+        return calculateFileValue(this.context, object, tag);
+    }
+
+    private _getFileTags(id: string): FileTags {
+        if (this._fileMap.has(id)) {
+            return this._fileMap.get(id);
+        }
+        const tags = {};
+        this._fileMap.set(id, tags);
+        return tags;
+    }
+}
