+++ src/aux-common/Files/index.ts
@@ -1,4 +1,5 @@
 export * from './File';
-export * from './FilesChannel';
+export * from './FileContext';
 export * from './FileCalculations';
+export * from './FilesChannel';
 export * from './TagUpdatedEvent';
